Funcion para guardar personajes y evitar errores al introducir datos (En Proeso)

+++ src/Bando/Marine.java
@@ -5,7 +5,7 @@ import Personajes.Personaje;
 public class Marine extends Personaje {
     protected String rango;
 
-    public Marine (String nombre, String atributo,int ataque, int salud,int defensa, int velocidad, int recompensa, String rango){
+    public Marine (String nombre, String atributo,int ataque, int salud,int defensa, int velocidad, long recompensa, String rango){
         super (nombre, atributo, ataque, salud, defensa, velocidad, recompensa);
         this.rango = rango;
     }
